# Optimization: Use std::map for tracking zero-copy buffers