--- a/libdeno/internal.h
+++ b/libdeno/internal.h
@@ -2,6 +2,7 @@
 #ifndef INTERNAL_H_
 #define INTERNAL_H_
 
+#include <map>
 #include <string>
 #include "deno.h"
 #include "third_party/v8/include/v8.h"
@@ -22,7 +23,7 @@ struct deno_s {
 
   int32_t pending_promise_events;
   v8::Persistent<v8::Context> context;
-  v8::Persistent<v8::Map> async_data_map;
+  std::map<int32_t, v8::Persistent<v8::Value>*> async_data_map;
   deno_recv_cb cb;
   int32_t next_req_id;
   void* user_data;